+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking?:** Resolve default manifest path and default package using `cargo metadata`
 - Improve error messages
 - Improve help output with sections
+- Add `--bin` argument to support `bin` targets
 
 ## [0.8.0] - 2025-07-18